# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Book API
  version: 1.0.0
  description: Sample CRUD API for books
servers:
  - url: /

tags:
  - name: book
    description: Book management

paths:
  /books:
    get:
      tags: [ book ]
      operationId: getAllBooks
      summary: Get all books
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      tags: [ book ]
      operationId: createBook
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

  /books/{id}:
    get:
      tags: [ book ]
      operationId: getBookById
      summary: Get book by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not Found
    put:
      tags: [ book ]
      operationId: updateBook
      summary: Update an existing book
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not Found
    delete:
      tags: [ book ]
      operationId: deleteBook
      summary: Delete a book by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found

components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    Book:
      type: object
      required: [ id, title, author ]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        year:
          type: integer
          format: int32

    CreateBookRequest:
      type: object
      required: [ title, author ]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        author:
          type: string
          minLength: 1
          maxLength: 255
        year:
          type: integer
          format: int32
          minimum: 0
          maximum: 9999

    UpdateBookRequest:
      type: object
      required: [ title, author ]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        author:
          type: string
          minLength: 1
          maxLength: 255
        year:
          type: integer
          format: int32
          minimum: 0
          maximum: 9999